+++ src/mml_utils/parse/parser.py
@@ -6,18 +6,18 @@ from mml_utils.parse.mmi import extract_mml_from_mmi_data
 from mml_utils.parse.xmi import extract_mml_from_xmi_data
 
 
-def extract_mml_data(file: pathlib.Path, *, encoding='cp1252', target_cuis=None, output_format='json',
+def extract_mml_data(file: pathlib.Path, *, encoding='cp1252', target_cuis=None, extract_format='json',
                      data_directories=None):
     with open(file, encoding=encoding) as fh:
         text = fh.read()
     if not text.strip():  # handle empty note
         return
-    if output_format == 'json':  # TODO: match-case
+    if extract_format == 'json':  # TODO: match-case
         data = json.loads(text)
         yield from extract_mml_from_json_data(data, file.name, target_cuis=target_cuis)
-    elif output_format == 'mmi':
+    elif extract_format == 'mmi':
         yield from extract_mml_from_mmi_data(text, file.name, target_cuis=target_cuis)
-    elif output_format == 'xmi':
+    elif extract_format == 'xmi':
         yield from extract_mml_from_xmi_data(text, file.name, target_cuis=target_cuis)
     else:
-        raise ValueError(f'Unrecognized output format: {output_format}.')
+        raise ValueError(f'Unrecognized output format: {extract_format}.')